requirements.txt
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.1
packaging==23.2
pluggy==1.4.0
pytest==8.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.27
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
